+++ savedata/save.json
@@ -2,7 +2,7 @@
     "entities": [
         {
             "Position": {
-                "x": 202.43760681152344,
+                "x": 364.1578674316406,
                 "y": 300.0
             },
             "id": 0
